+++ src/AngleSharp.Css/Dom/ScrollSnapAxis.cs
@@ -1,0 +1,29 @@
+namespace AngleSharp.Css.Dom
+{
+    /// <summary>
+    /// An enumeration with all possible scroll snap axis options.
+    /// </summary>
+    public enum ScrollSnapAxis : byte
+    {
+        /// <summary>
+        /// The scroll container snaps to snap positions in its horizontal axis only
+        /// </summary>
+        X,
+        /// <summary>
+        /// The scroll container snaps to snap positions in its horizontal axis only
+        /// </summary>
+        Y,
+        /// <summary>
+        /// The scroll container snaps to snap positions in its block axis only
+        /// </summary>
+        Block,
+        /// <summary>
+        /// The scroll container snaps to snap positions in its inline axis only
+        /// </summary>
+        Inline,
+        /// <summary>
+        /// The scroll container snaps to snap positions in its inline axis only
+        /// </summary>
+        Both,
+    }
+}